+++ CopilotKit/packages/react-core/src/types/index.ts
@@ -1,4 +1,8 @@
 export type { DocumentPointer } from "./document-pointer";
 export type { SystemMessageFunction } from "./system-message";
-export type { ActionRenderProps, RenderFunctionStatus } from "./frontend-action";
+export type {
+  ActionRenderProps,
+  RenderFunctionStatus,
+  CatchAllActionRenderProps,
+} from "./frontend-action";
 export type { CopilotChatSuggestionConfiguration } from "./chat-suggestion-configuration";