+++ examples/coagents-research-canvas/agent/research_canvas/demo.py
@@ -9,7 +9,6 @@ from fastapi import FastAPI
 import uvicorn
 from copilotkit.integrations.fastapi import add_fastapi_endpoint
 from copilotkit import CopilotKitSDK, LangGraphAgent
-from copilotkit.langchain import copilotkit_messages_to_langchain
 from research_canvas.agent import graph
 
 app = FastAPI()
@@ -40,4 +39,16 @@ def health():
 def main():
     """Run the uvicorn server."""
     port = int(os.getenv("PORT", "8000"))
-    uvicorn.run("research_canvas.demo:app", host="0.0.0.0", port=port)
+    uvicorn.run(
+        "research_canvas.demo:app",
+        host="0.0.0.0",
+        port=port,
+        reload=True,
+        reload_dirs=(
+            ["."] +
+            (["../../../sdk-python/copilotkit"]
+             if os.path.exists("../../../sdk-python/copilotkit")
+             else []
+             )
+        )
+    )