+++ CopilotKit/packages/runtime-client-gql/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @copilotkit/runtime-client-gql
 
+## 1.5.0
+
+### Minor Changes
+
+- 1b47092: Synchronize LangGraph messages with CopilotKit
+
+### Patch Changes
+
+- 1b47092: CoAgents v0.3 prerelease
+- Updated dependencies [1b47092]
+- Updated dependencies [1b47092]
+  - @copilotkit/shared@1.5.0
+
+## 1.5.0-coagents-v0-3.0
+
+### Minor Changes
+
+- Synchronize LangGraph messages with CopilotKit
+
+### Patch Changes
+
+- e66bce4: CoAgents v0.3 prerelease
+- Updated dependencies
+- Updated dependencies [e66bce4]
+  - @copilotkit/shared@1.5.0-coagents-v0-3.0
+
 ## 1.4.8
 
 ### Patch Changes