+++ CopilotKit/packages/shared/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @copilotkit/shared
 
+## 1.5.0
+
+### Minor Changes
+
+- 1b47092: Synchronize LangGraph messages with CopilotKit
+
+### Patch Changes
+
+- 1b47092: CoAgents v0.3 prerelease
+
+## 1.5.0-coagents-v0-3.0
+
+### Minor Changes
+
+- Synchronize LangGraph messages with CopilotKit
+
+### Patch Changes
+
+- e66bce4: CoAgents v0.3 prerelease
+
 ## 1.4.8
 
 ### Patch Changes